requirements.txt
Flask>=2.2.2
gymnasium>=0.28.1
langchain>=0.0.225
openai>=0.27.7
matplotlib>=3.7.1
numpy>=1.24.3
PyYAML>=6.0
rich
moviepy